refactor: 패키지 이름 변경 및 이동
