+++ src/main/java/ChatManagement/chat/repository/ChatMessageRepository.java
@@ -1,8 +1,6 @@
 package ChatManagement.chat.repository;
 
-import ChatManagement.chat.dao.ChatMessage;
-import ChatManagement.chat.dao.ChatRoom;
-import java.util.List;
+import ChatManagement.chat.domain.ChatMessage;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
