+++ src/main/java/ChatManagement/kafka/service/dto/type/LogType.java
@@ -1,0 +1,5 @@
+package ChatManagement.kafka.service.dto.type;
+
+public enum LogType {
+    INFO, ERROR
+}
