--- src/main/java/ChatManagement/kafka/type/LogType.java
@@ -1,5 +1,0 @@
-package ChatManagement.kafka.type;
-
-public enum LogType {
-    INFO, ERROR
-}
